refactor(project-monitoring): Change auth middleware ref
